requirements.txt
supabase>=2.0.0
perchance>=0.1.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
Pillow>=10.0.0
